+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.2] 2022-01-31
+## [Unreleased]
 ### Added
 - Added support for a custom port number to be set.
+- Added a new `vitamin:serve` Artisan command to replace the direct call to `npx vite serve`.
 - Change to an HTTP HEAD request instead of a GET request when checking if the dev server is running.
 
 ## [0.3.1] 2022-01-27